Conditional imports for face detection
